Add condition on script to check the jobs list file

+++ tiler-server/seed-by-diffs.sh
@@ -7,32 +7,17 @@ workDir=/mnt/data
 expire_dir=$workDir/imposm/imposm3_expire_dir
 mkdir -p $expire_dir
 
-# this will be essentially treated as a pidfile
+# This will be essentially treated as a pidfile
 queued_jobs=$workDir/imposm/in_progress.list
-# output of seeded
+# Output of seeded
 completed_jobs=$workDir/imposm/completed.list
 
-# a directory to place the worked expiry lists
-# THIS MUST BE OUTSIDE OF $expire_dir
+# Directory to place the worked expiry lists
 completed_dir=$workDir/imposm/imposm3_expire_purged
 mkdir -p $completed_dir
 
-# assert no other jobs are running
-# if [[ -f $queued_jobs ]]; then
-#   exit
-# else
-#   touch $queued_jobs
-#   if [[ ! $? ]]; then
-#     rm $queued_jobs
-#     exit
-#   fi
-# fi
 
-
-# files newer than this amount of seconds will
-# not be used for this job
-
-# list Files in dir
+# List files in expire_dir
 imp_list=`find $expire_dir -type f`
 
 for f in $imp_list; do
@@ -40,11 +25,13 @@ for f in $imp_list; do
 done
 
 # Sort the files and set unique rows
-sort -u $queued_jobs > $workDir/imposm/tmp.list && mv $workDir/imposm/tmp.list $queued_jobs
+if [ -f $queued_jobs ] ; then
+    sort -u $queued_jobs > $workDir/imposm/tmp.list && mv $workDir/imposm/tmp.list $queued_jobs
+fi
 
 for f in $imp_list; do
     echo "seeding from $f"
-    # Read each line of the tiles file
+    # Read each line on the tiles file
     while IFS= read -r tile
     do
         bounds="$(python tile2bounds.py $tile)"
@@ -71,5 +58,7 @@ for f in $imp_list; do
     mv $f $completed_dir
 done
 
-echo "finished seeding"
-rm $queued_jobs
+if [ -f $queued_jobs ] ; then
+    echo "finished seeding"
+    rm $queued_jobs
+fi
